Unwrap predeclared Copy scalars at function return boundaries
Functions whose Go signature returns a single predeclared Copy scalar
(int family, uint family, float32/64, bool, uintptr) now emit a bare
Rust scalar in the return type instead of Rc<RefCell<Option<T>>>. The
canonical case from external feedback:

    func GetYear() int { return 2024 }

now lowers to:

    pub fn get_year() -> i32 { return 2024 as i32; }

instead of wrapping the result in Rc<RefCell<Option<i32>>>. There is no
addressable callee return slot for a non-pointer scalar, so the wrapper
allocation had no semantic effect — it was pure overhead.

Scope:
- Single-result functions emit bare scalar return types.
- Multi-result tuple elements also unwrap scalar positions via the
  GoReturnTypeToRust helper, with named-return paths updated.
- Caller sites stop applying .borrow().as_ref().unwrap() to bare scalar
  return values; storage into a wrapped local re-wraps once.
- Predeclared builtins (len, cap, copy, ...) and stdlib calls keep
  their existing lowering — callReturnsBareScalar is gated to
  user-defined functions in the current package.

Test mechanism:
- tests.bats: when a fixture has expected_main.rs, the freshly
  transpiled main.rs is diffed against it and the test fails on any
  mismatch. This pins the generated Rust shape for fixtures that
  document a specific lowering decision.
- tests/scalar_return_unwrapped/: new fixture with the GetYear example
  and a pinned expected_main.rs locking in the bare i32 return.

XFAIL tests/source_stdlib_path_filepath_isabs/ auto-promotes from this
change: filepath.IsAbs returning bool now lowers cleanly through the
source-stdlib pipeline.

The multi-result expansion is still rough around the edges: a handful
of destructuring sites (range-var indices, stdlib comma-ok results
flowing through scalar return positions) still produce type
mismatches. The single-result path the external critique called out is
fully working.

+++ tests/cross_file_lower_const_compound/flags.rs
@@ -1,1 +1,1 @@
-pub(crate) const FLAG_SYNC_MARKERS: i64 = 1 << 0;
+pub(crate) const FLAG_SYNC_MARKERS: i32 = 1 << 0;
